Keep partition child lock until end of select
Introduce a new GUC gp_keep_partition_children_locks that preserves
relation locks on AO partition leaves, that are taken during planning.

This fixes concurrency issues such as #16545, where a SELECT pre-dating
a concurrent lazy vacuum can read past eof of a segfile, since the
vacuum would "drop" the segment. Adequate locking during planning is one
way to ensure that vacuumStatement_Relation() can't acquire the
AccessExclusiveLock on the coordinator. This would cause it to give up
on the drop, and the concurrent SELECT isn't compromised.

Unfortunately, we can't quite mark this GUC on by default, as it has
consequences on the amount of locks consumed:

Setting this GUC on means that we keep the acquired locks in
expand_inherited_rtentry() around longer now, up until the end of the
transaction. This can have consequences for transactions dealing with a
large number of partitions, leading us to inch closer to the
max_locks_per_transaction limit. Joins/unions etc with large partition
tables now consume more locks than before.

Co-authored-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>
Co-authored-by: Ashwin Agrawal <aashwin@vmware.com>